Split library packaging locale tests
